+++ dao/spirit_dao.go
@@ -12,20 +12,28 @@ const (
 
 // SpiritDAO is the DAO interface to work with spirits
 type SpiritDAO interface {
-	//
+
+	// GetSpiritByID returns a spirit by its ID
 	GetSpiritByID(ID string) (*model.Spirit, error)
-	//
-	GetAllSpirits(skip, limit int) ([]model.Spirit, error)
-	//
+
+	// GetAllSpirits returns all spirits with paging capability
+	GetAllSpirits(start, end int) ([]model.Spirit, error)
+
+	// GetSpiritsByName returns all spirits by name
 	GetSpiritsByName(name string) ([]model.Spirit, error)
-	//
+
+	// GetSpiritsByType returns all spirits by type
 	GetSpiritsByType(spiritType string) ([]model.Spirit, error)
-	//
+
+	// GetSpiritsByTypeAndScore returns all spirits by type and score greater than parameter
 	GetSpiritsByTypeAndScore(spiritType string, score uint8) ([]model.Spirit, error)
-	//
+
+	// SaveSpirit saves the spirit
 	SaveSpirit(spirit *model.Spirit) error
-	//
+
+	// UpsertSpirit updates or creates a spirit
 	UpsertSpirit(ID string, spirit *model.Spirit) (*mgo.ChangeInfo, error)
-	//
+
+	// DeleteSpirit deletes a spirits by its ID
 	DeleteSpirit(ID string) error
 }