+++ utils/web.go
@@ -21,25 +21,27 @@ const (
 func SendJSONWithHTTPCode(w http.ResponseWriter, d interface{}, code int) {
 	w.Header().Set(ResponseHeaderContentTypeKey, ResponseHeaderContentTypeJSONUTF8)
 	w.WriteHeader(code)
-	err := json.NewEncoder(w).Encode(d)
-	if err != nil {
-		logger.WithField("body", d).WithField("code", code).Error("error while encoding JSON body of response")
-		// panic will cause the http.StatusInternalServerError to be send to users
-		panic(err)
+	if d != nil {
+		err := json.NewEncoder(w).Encode(d)
+		if err != nil {
+			logger.WithField("body", d).WithField("code", code).Error("error while encoding JSON body of response")
+			// panic will cause the http.StatusInternalServerError to be send to users
+			panic(err)
+		}
 	}
 }
 
-// SendJSON outputs a JSON with http.StatusOK code
+// SendJSONOk outputs a JSON with http.StatusOK code
 func SendJSONOk(w http.ResponseWriter, d interface{}) {
 	SendJSONWithHTTPCode(w, d, http.StatusOK)
 }
 
-// SendError sends error with a custom message and error code
+// SendJSONError sends error with a custom message and error code
 func SendJSONError(w http.ResponseWriter, error string, code int) {
 	SendJSONWithHTTPCode(w, map[string]string{errorMsg: error}, code)
 }
 
-// SendNotFound produces a http.StatusNotFound response with the following JSON, '{"Error":"Resource not found"}'
+// SendJSONNotFound produces a http.StatusNotFound response with the following JSON, '{"Error":"Resource not found"}'
 func SendJSONNotFound(w http.ResponseWriter) {
 	SendJSONError(w, resourceNotFound, http.StatusNotFound)
 }
@@ -56,4 +58,10 @@ func ParamAsString(name string, r *http.Request) string {
 	vars := mux.Vars(r)
 	value := vars[name]
 	return value
+}
+
+// GetJSONContent returns the JSON content of a request
+func GetJSONContent(v interface{}, r *http.Request) error {
+	defer r.Body.Close()
+	return json.NewDecoder(r.Body).Decode(v)
 }